# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: CDT Jamaica Management API with Odoo Community integration
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /students:
    get:
      operationId: listStudents
      tags: [students]
      summary: List all students
      parameters:
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: classId
          schema:
            type: ["integer", "null"]
        - in: query
          name: status
          schema:
            type: ["string", "null"]
            enum: [active, inactive, null]
      responses:
        "200":
          description: List of students
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Student"
    post:
      operationId: createStudent
      tags: [students]
      summary: Legacy student creation endpoint; linked records are managed in Odoo
      deprecated: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudentInput"
      responses:
        "201":
          description: Created student
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
        "409":
          description: Student and enrollment changes must be made in Odoo
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /students/{id}:
    get:
      operationId: getStudent
      tags: [students]
      summary: Get a student by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Student
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateStudent
      tags: [students]
      summary: Legacy student update endpoint; linked records are managed in Odoo
      deprecated: true
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudentUpdate"
      responses:
        "200":
          description: Updated student
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
        "409":
          description: Student and enrollment changes must be made in Odoo
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteStudent
      tags: [students]
      summary: Legacy student deletion endpoint; linked records are managed in Odoo
      deprecated: true
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "409":
          description: Student and enrollment changes must be made in Odoo
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Classes ────────────────────────────────────────────────────────────────
...
  /finances/invoices:
    get:
      operationId: listInvoices
      tags: [finances]
      summary: List invoices from Odoo
      parameters:
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: state
          schema:
            type: ["string", "null"]
            enum: [draft, posted, cancelled, null]
      responses:
        "200":
          description: List of invoices
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Invoice"
...
  /performances:
    get:
      operationId: listPerformances
      tags: [performances]
      summary: List performances
      parameters:
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: upcoming
          schema:
            type: boolean
      responses:
        "200":
          description: List of performances
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Performance"
    post:
      operationId: createPerformance
      tags: [performances]
      summary: Create a performance
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PerformanceInput"
      responses:
        "201":
          description: Created performance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Performance"
  /performances/{id}:
    get:
      operationId: getPerformance
      tags: [performances]
      summary: Get a performance by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Performance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Performance"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updatePerformance
      tags: [performances]
      summary: Update a performance
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PerformanceUpdate"
      responses:
        "200":
          description: Updated performance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Performance"
        "409":
          description: Sanity-managed performance fields are read-only until cutover
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deletePerformance
      tags: [performances]
      summary: Delete a performance
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "409":
          description: Sanity-managed performances cannot be deleted until cutover
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Tasks (Odoo) ────────────────────────────────────────────────────────
...
  /reports/finances:
    get:
      operationId: getFinanceReport
      tags: [reports]
      summary: Monthly revenue breakdown
      parameters:
        - in: query
          name: year
          schema:
            type: integer
      responses:
        "200":
          description: Finance report
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MonthlyRevenue"

  # ── Website content mirror ───────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    Student:
      type: object
      properties:
        id:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        dateOfBirth:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, inactive]
        enrolledClassIds:
          type: array
          items:
            type: integer
        guardianName:
          type: ["string", "null"]
        guardianPhone:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        studentNumber:
          type: string
        contactId:
          type: ["integer", "null"]
        contactName:
          type: string
        academicStatus:
          type: string
        classNames:
          type: array
          items:
            type: string
        programName:
          type: ["string", "null"]
        tuitionBilled:
          type: number
        tuitionPaid:
          type: number
        tuitionBalance:
          type: number
        tuitionStatus:
          type: ["string", "null"]
          enum: [draft, due, partial, paid, cancelled, null]
        currency:
          type: string
        sampleData:
          type: boolean
      required: [id, firstName, lastName, status, enrolledClassIds, createdAt, studentNumber, contactId, contactName, academicStatus, classNames, programName, tuitionBilled, tuitionPaid, tuitionBalance, tuitionStatus, currency, sampleData]
    StudentInput:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
        dateOfBirth:
          type: string
        status:
          type: string
          enum: [active, inactive]
        enrolledClassIds:
          type: array
          items:
            type: integer
        guardianName:
          type: string
        guardianPhone:
          type: string
        notes:
          type: string
      required: [firstName, lastName]
    StudentUpdate:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
        dateOfBirth:
          type: string
        status:
          type: string
          enum: [active, inactive]
        enrolledClassIds:
          type: array
          items:
            type: integer
        guardianName:
          type: string
        guardianPhone:
          type: string
        notes:
          type: string

    # ── DanceClass ─────────────────────────────────────────────────────────
...
    Invoice:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        partnerName:
          type: string
        amountTotal:
          type: number
        amountDue:
          type: number
        currency:
          type: string
        invoiceDate:
          type: ["string", "null"]
        dueDate:
          type: ["string", "null"]
        state:
          type: string
          enum: [draft, posted, cancelled]
        paymentState:
          type: string
          enum: [not_paid, in_payment, paid, partial, reversed, null]
        odooId:
          type: ["string", "null"]
      required: [id, name, partnerName, amountTotal, amountDue, currency, state, paymentState]
...
    Performance:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        date:
          type: string
        time:
          type: ["string", "null"]
        venue:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [upcoming, completed, cancelled]
        participatingClassIds:
          type: array
          items:
            type: integer
        ticketPrice:
          type: ["number", "null"]
        capacity:
          type: ["integer", "null"]
        createdAt:
          type: string
        managedBySanity:
          type: boolean
      required: [id, title, date, venue, status, participatingClassIds, createdAt, managedBySanity]
    PerformanceInput:
      type: object
      properties:
        title:
          type: string
        date:
          type: string
        time:
          type: string
        venue:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [upcoming, completed, cancelled]
        participatingClassIds:
          type: array
          items:
            type: integer
        ticketPrice:
          type: number
        capacity:
          type: integer
      required: [title, date, venue]
    PerformanceUpdate:
      type: object
      properties:
        title:
          type: string
        date:
          type: string
        time:
          type: string
        venue:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [upcoming, completed, cancelled]
        participatingClassIds:
          type: array
          items:
            type: integer
        ticketPrice:
          type: number
        capacity:
          type: integer

    # ── Task ───────────────────────────────────────────────────────────────
...
    MonthlyRevenue:
      type: object
      properties:
        month:
          type: string
        revenue:
          type: number
        invoiceCount:
          type: integer
        donations:
          type: number
      required: [month, revenue, invoiceCount, donations]