# lib/api-zod/src/generated/types/monthlyRevenue.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * CDT Jamaica Management API with Odoo Community integration
 * OpenAPI spec version: 0.1.0
 */

export interface MonthlyRevenue {
  month: string;
  revenue: number;
  invoiceCount: number;
  donations: number;
}